fix: prevent `unfold` from dropping samples and trim the reconstructed audio
`unfold` discards samples at the end of a tensor if the padded length is
not perfectly divisible by `hop_size`. When `overlap_ratio` is low, `hop_size`
is large so the likelihood of non-zero remainders (and misalignment) increases.

we use constant padding to prevent unfold from dropping frames, and more
robustly trim off the initial symmetric padding.